+++ test/conftest.py
@@ -1,0 +1,7 @@
+# pytest-embedded configuration
+import pytest
+
+@pytest.fixture(scope='session')
+def app_path():
+    return '.'
+